Add reading/bookmark DTOs, endpoints & repo queries
Introduce user-centric reading/bookmark features and related optimizations:

- Add DTOs: BookmarkDTO, ReadingHistoryDTO, ReadingActivityDTO, UserStatsDTO.
- Extend ArticleController with endpoints for reading-history, reading-activity, bookmarks, bookmark folders and user-stats (with auth checks).
- Expand ArticleService interface and implement new methods in ArticleServiceImpl: getReadingHistory, getReadingActivity, getBookmarks, getBookmarksByFolder, getBookmarkFolders, getUserStats; plus DTO mappers.
- Add repository query methods to efficiently batch-load engagement data: ArticleLikeRepository.findLikedArticleIds, BookmarkRepository.findBookmarkedArticleIds & paginated folder query, ReadingHistoryRepository.findReadArticleIds and extra analytics queries; change NotificationPreferenceRepository ID type to Long.
- Replace previous in-memory/filtering approach in toPageDTO with targeted repo queries to avoid N+1 and full-table scans.
- Move several services from package service.impl to serviceimpl and update imports (NewsIngestionService, NotificationServiceImpl, UserServiceImpl, ArticleServiceImpl). Adjust various imports/usages accordingly.
- Minor cleanup: remove unused annotations/imports, comment out a notification permit in SecurityConfig, and initialize default notification preferences when registering a user.

These changes add new user-facing APIs for reading/bookmarks, improve performance by using targeted queries, and reorganize service packages.

+++ src/main/java/com/news/digest/app/repository/NotificationPreferenceRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface NotificationPreferenceRepository  extends JpaRepository<NotificationPreference, Integer> {
+public interface NotificationPreferenceRepository  extends JpaRepository<NotificationPreference, Long> {
 
     List<NotificationPreference> findByUserId(Long userId);
